# Conflicts:
#	Softeq.NetKit.Chat.Web/Startup.cs

+++ Softeq.NetKit.Chat.Domain.Services/DomainServices/MemberService.cs
@@ -10,7 +10,6 @@ using EnsureThat;
 using Softeq.NetKit.Chat.Data.Persistent;
 using Softeq.NetKit.Chat.Domain.DomainModels;
 using Softeq.NetKit.Chat.Domain.Exceptions;
-using Softeq.NetKit.Chat.Domain.Exceptions.ErrorHandling;
 using Softeq.NetKit.Chat.Domain.Services.Configuration;
 using Softeq.NetKit.Chat.Domain.Services.Mappers;
 using Softeq.NetKit.Chat.Domain.TransportModels.Request.Channel;
@@ -26,38 +25,35 @@ namespace Softeq.NetKit.Chat.Domain.Services.DomainServices
     {
         private readonly CloudStorageConfiguration _configuration;
 
-        public MemberService(IUnitOfWork unitOfWork, CloudStorageConfiguration configuration) : base(unitOfWork)
+        public MemberService(IUnitOfWork unitOfWork, CloudStorageConfiguration configuration)
+            : base(unitOfWork)
         {
+            Ensure.That(configuration).IsNotNull();
+
             _configuration = configuration;
-        }
-        
-        public async Task<ParticipantResponse> SurelyGetMemberBySaasUserIdAsync(string saasUserId)
-        {
-            var member = await UnitOfWork.MemberRepository.GetMemberBySaasUserIdAsync(saasUserId);
-            Ensure.That(member).WithException(x => new ServiceException(new ErrorDto(ErrorCode.NotFound, "Member does not exist."))).IsNotNull();
-            return member.ToParticipantResponse();
         }
 
         //TODO: Add Unit Tests
         public async Task<MemberSummary> GetMemberBySaasUserIdAsync(string saasUserId)
         {
             var member = await UnitOfWork.MemberRepository.GetMemberBySaasUserIdAsync(saasUserId);
-            Ensure.That(member).WithException(x => new ServiceException(new ErrorDto(ErrorCode.NotFound, "Member does not exist."))).IsNotNull();
+            if (member == null)
+            {
+                throw new NetKitChatNotFoundException($"Unable to get member by {nameof(saasUserId)}. Member {nameof(saasUserId)}:{saasUserId} not found.");
+            }
+
             return member.ToMemberSummary(_configuration);
         }
 
         public async Task<MemberSummary> GetMemberByIdAsync(Guid memberId)
         {
             var member = await UnitOfWork.MemberRepository.GetMemberByIdAsync(memberId);
-            Ensure.That(member).WithException(x => new ServiceException(new ErrorDto(ErrorCode.NotFound, "Member does not exist."))).IsNotNull();
-            return member.ToMemberSummary(_configuration);
-        }
+            if (member == null)
+            {
+                throw new NetKitChatNotFoundException($"Unable to get member by {nameof(memberId)}. Member {nameof(memberId)}:{memberId} not found.");
+            }
 
-        public async Task<ParticipantResponse> GetMemberAsync(UserRequest request)
-        {
-            var member = await UnitOfWork.MemberRepository.GetMemberBySaasUserIdAsync(request.SaasUserId);
-            Ensure.That(member).WithException(x => new ServiceException(new ErrorDto(ErrorCode.NotFound, "Member does not exist."))).IsNotNull();
-            return member.ToParticipantResponse();
+            return member.ToMemberSummary(_configuration);
         }
 
         public async Task<IReadOnlyCollection<MemberSummary>> GetChannelMembersAsync(Guid channelId)
@@ -69,15 +65,21 @@ namespace Softeq.NetKit.Chat.Domain.Services.DomainServices
         public async Task<ChannelResponse> InviteMemberAsync(InviteMemberRequest request)
         {
             var channel = await UnitOfWork.ChannelRepository.GetChannelByIdAsync(request.ChannelId);
-            Ensure.That(channel).WithException(x => new NotFoundException(new ErrorDto(ErrorCode.NotFound, "Channel does not exist."))).IsNotNull();
+            if (channel == null)
+            {
+                throw new NetKitChatNotFoundException($"Unable to invite member. Channel {nameof(request.ChannelId)}:{request.ChannelId} not found.");
+            }
+
             var member = await UnitOfWork.MemberRepository.GetMemberByIdAsync(request.MemberId);
-            Ensure.That(member).WithException(x => new ServiceException(new ErrorDto(ErrorCode.NotFound, "Member does not exist."))).IsNotNull();
+            if (member == null)
+            {
+                throw new NetKitChatNotFoundException($"Unable to invite member. Member {nameof(request.MemberId)}:{request.MemberId} not found.");
+            }
 
             var channelMembers = await GetChannelMembersAsync(channel.Id);
-
             if (channelMembers.Any(x => x.Id == request.MemberId))
             {
-                throw new ConflictException(new ErrorDto(ErrorCode.ConflictError, "User already exists in channel."));
+                throw new NetKitChatInvalidOperationException($"Unable to invite member. Member {nameof(request.MemberId)}:{request.MemberId} already joined channel {nameof(request.ChannelId)}:{request.ChannelId}.");
             }
 
             var channelMember = new ChannelMembers
@@ -98,12 +100,6 @@ namespace Softeq.NetKit.Chat.Domain.Services.DomainServices
             var newChannel = await UnitOfWork.ChannelRepository.GetChannelByIdAsync(channel.Id);
 
             return newChannel.ToChannelResponse(_configuration);
-        }
-
-        public async Task<IReadOnlyCollection<ParticipantResponse>> GetOnlineChannelMembersAsync(ChannelRequest request)
-        {
-            var members = await UnitOfWork.MemberRepository.GetOnlineMembersInChannelAsync(request.ChannelId);
-            return members.Select(x => x.ToParticipantResponse()).ToList().AsReadOnly();
         }
 
         public async Task<ClientResponse> GetOrAddClientAsync(AddClientRequest request)
@@ -155,7 +151,11 @@ namespace Softeq.NetKit.Chat.Domain.Services.DomainServices
         public async Task DeleteClientAsync(DeleteClientRequest request)
         {
             var client = await UnitOfWork.ClientRepository.GetClientByConnectionIdAsync(request.ClientConnectionId);
-            Ensure.That(client).WithException(x => new NotFoundException(new ErrorDto(ErrorCode.NotFound, "Client does not exist.")));
+            if (client == null)
+            {
+                throw new NetKitChatNotFoundException($"Unable to delete client. Client {nameof(request.ClientConnectionId)}:{request.ClientConnectionId} not found.");
+            }
+
             await UnitOfWork.ClientRepository.DeleteClientAsync(client.Id);
 
             var clients = await GetMemberClientsAsync(client.MemberId);
@@ -177,6 +177,7 @@ namespace Softeq.NetKit.Chat.Domain.Services.DomainServices
             var member = await UnitOfWork.MemberRepository.GetMemberBySaasUserIdAsync(saasUserId);
             if (member != null)
             {
+                // TODO [az]: Should we throw exception here?
                 return member.ToMemberSummary(_configuration);
             }
 
@@ -199,10 +200,15 @@ namespace Softeq.NetKit.Chat.Domain.Services.DomainServices
 
         public async Task UpdateMemberStatusAsync(UpdateMemberStatusRequest request)
         {
-            var member = await SurelyGetMemberBySaasUserIdAsync(request.SaasUserId);
+            var member = await UnitOfWork.MemberRepository.GetMemberBySaasUserIdAsync(request.SaasUserId);
+            if (member == null)
+            {
+                throw new NetKitChatNotFoundException($"Unable to update member status. Member {nameof(request.SaasUserId)}:{request.SaasUserId} not found.");
+            }
+
             member.Status = request.UserStatus;
             member.LastActivity = DateTimeOffset.Now;
-            await UnitOfWork.MemberRepository.UpdateMemberAsync(member.ToMember());
+            await UnitOfWork.MemberRepository.UpdateMemberAsync(member);
         }
 
         public async Task<IReadOnlyCollection<ClientResponse>> GetClientsByMemberIds(List<Guid> memberIds)
